shaders: Modify ucompare statement template to work around a Qualcomm shader compiler bug.
The actual problem was the usage of certain functions on the guest
side in a virtio-gpu setup, where the shader code contained the step()
function. After translation to the internal TGSI shader format, it
will be unpacked by virglrenderer into GLSL again, and the template in
use created a construct where the Qualcomm shader compiler had a
problem with.

The template originally creates constructs like this:

  float val = float(uintBitsToFloat(uint(greaterThanEqual(tmpA, vec4(0.8)).x) * uint(0xffffffff)));

which the new template generates as:

  float val = float(greaterThanEqual(tmpA, vec4(0.8)).x);
  val = float(uintBitsToFloat(uint(val) * uint(0xffffffff)));

This is (almost) the same, but the former doesn't work on Qualcomm
systems, while the latter is running without problems.

Issue: COQOS-8846, E0910-83
Change-Id: I590ef29c86bbb370c651cb18b39a268c62cb58f3