shader: Avoid requiring GL_EXT_texture_shadow_lod if not supported.
This patch ensures that the 'GL_EXT_texture_shadow_lod' extension is
not required or appended in the shader header when it is not supported
by the host. This prevents potential shader compilation issues on
hosts that lack support for this extension, while maintaining correct
behavior for features around GLSL sampling that do not require this
extension.

Issue: COQOS-19616
Change-Id: I698af8b8b11b8d0c8f764fcf28420ff899fb17e8