vrend: Disable dual_src_blend feature (GL_EXT_blend_func_extended) for Qualcomm.
The dual_src_blend feature as in support for GL_EXT_blend_func_extended
seems to be broken on Qualcomm Adreno. Reports from the past about broken
MRT (multi render target) support can be found on the internet as well,
so it looks like that didn't get fixed, yet.

The error messages happen at the linking stage, and follow a pattern similar
to this: "Error: Output ... location or component exceeds max allowed."
where the output variable is part of the multiple outputs from the fragment
shader stage.

With this change the feature is suppressed even if the driver reports
GL_EXT_blend_func_extended being available. Being an optional feature,
virglrenderer will adapt accordingly. This fix can also be used on
other platforms, however, it may limit their capabilities this way.
Exynos/Mali doesn't support this extension, for example, so it is safe
to use there.

Note: This is a workaround for a OpenGL/ES driver issue.

Issue: COQOS-8161
Change-Id: Id5e2a37222fabf886c1263c39305fd05236a95da